fix: align oldarch Java Privacy spec with newarch
newarch Codegen specs use `double` for all JS `number` types, including
enums.  This allows the PrivacyManager implementation to align to both
new and old architectures.

+++ android/src/oldarch/com/gimbal/rtn/PrivacyManagerSpec.java
@@ -17,9 +17,9 @@ abstract class PrivacyManagerSpec extends ReactContextBaseJavaModule {
 
     public abstract void getGdprConsentRequirement(Promise promise);
 
-    public abstract void setUserConsent(int consentType, int state);
+    public abstract void setUserConsent(double consentType, double state);
 
-    public abstract void getUserConsent(int consentType, Promise promise);
+    public abstract void getUserConsent(double consentType, Promise promise);
 
     @Nullable
     public final Map<String, Object> getConstants() {
